+++ flash/core/data/batch.py
@@ -230,7 +230,7 @@ class _Preprocessor(torch.nn.Module):
                 with self._collate_context:
                     samples, metadata = self._extract_metadata(samples)
                     samples = self.collate_fn(samples)
-                    if metadata:
+                    if metadata and isinstance(samples, dict):
                         samples[DefaultDataKeys.METADATA] = metadata
                     self.callback.on_collate(samples, self.stage)